requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
dj-database-url==2.3.0
dj-email-url==1.0.6
Django==5.1.2
django-allauth==65.1.0
django-cache-url==3.4.5
django-crispy-forms==2.3
django-js-asset==2.2.0
django-mptt==0.16.0
environs==11.0.0
marshmallow==3.23.1
packaging==24.1
pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2021.5
